contracts linter and prettier

+++ contracts/ICommission.sol
@@ -1,5 +1,5 @@
 pragma solidity ^0.8.0;
 
 interface ICommission {
-    function increasePriority() external view returns (uint);
+    function increasePriority() external view returns (uint256);
 }
