+++ contracts/IReputation.sol
@@ -1,5 +1,9 @@
 pragma solidity ^0.8.0;
 
 interface IReputation {
-    function measure(uint createdAt, uint successRate, uint manualRate) external view returns (uint);
+    function measure(
+        uint256 createdAt,
+        uint256 successRate,
+        uint256 manualRate
+    ) external view returns (uint256);
 }
