+++ contracts/ISet.sol
@@ -1,7 +1,9 @@
 pragma solidity ^0.8.0;
 
 interface ISet {
-    function insert(uint) external;
-    function remove(uint) external;
-    function indices() external view returns( uint[] memory);
+    function insert(uint256) external;
+
+    function remove(uint256) external;
+
+    function indices() external view returns (uint256[] memory);
 }
