+++ contracts/Participants.sol
@@ -2,29 +2,42 @@ pragma solidity ^0.8.0;
 
 // todo: doers is an interface. court, doer, (customer?)... doer categories?
 contract Participants {
-
     // maybe Doer is a contract?
     struct Doer {
         address addr;
-        int level;
-        uint successOrders;
-        uint failedOrders;
-        uint registerDate;
+        int256 level;
+        uint256 successOrders;
+        uint256 failedOrders;
+        uint256 registerDate;
     }
 
-
-    mapping(address => uint) internal addrToIdx;
+    mapping(address => uint256) internal addrToIdx;
     Doer[] public doers;
 
     function newDoer(address _doer) external {
         if (doers.length > 0) {
-            require(doers[addrToIdx[_doer]].addr != _doer, "doer already exists");
+            require(
+                doers[addrToIdx[_doer]].addr != _doer,
+                "doer already exists"
+            );
         }
-        doers.push(Doer({addr: _doer, level: 0, successOrders: 0, failedOrders:0, registerDate: block.timestamp}));
-        addrToIdx[_doer] = doers.length-1;
+        doers.push(
+            Doer({
+                addr: _doer,
+                level: 0,
+                successOrders: 0,
+                failedOrders: 0,
+                registerDate: block.timestamp
+            })
+        );
+        addrToIdx[_doer] = doers.length - 1;
     }
 
-    function getDoerByAddress(address _addr) internal view returns (Doer storage) {
+    function getDoerByAddress(address _addr)
+        internal
+        view
+        returns (Doer storage)
+    {
         Doer storage d = doers[addrToIdx[_addr]];
         require(d.addr == _addr, "invalid doer");
         return d;
@@ -39,7 +52,7 @@ contract Participants {
         }
     }
 
-    function getSuccessRate() external view returns (uint) {
+    function getSuccessRate() external view returns (uint256) {
         // todo
         return 0;
     }
